+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+# v1.7.0-rc.1
+## mm/dd/2019
+
+1. [](#bugfix)
+    * Fixed `Page::untranslatedLanguages()` not being symmetrical to `Page::translatedLanguages()`
+
 # v1.7.0-beta.10
 ## 10/03/2019
 
-1. [](#new)
 1. [](#improved)
     * Flex: Removed extra exists check when creating object (messes up "non-existing" pages)
     * Support customizable null character replacement in `CSVFormatter::decode()`